requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
python-dotenv
pdfminer.six
chromadb
ipywidgets
Jinja2
ragas
datasets
matplotlib
seaborn
numpy